gradio_app/requirements.lock.txt
aiofiles==23.2.1
alembic==1.17.1
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
blinker==1.9.0
Bottleneck @ file:///home/task_176193795831906/conda-bld/bottleneck_1761938007449/work
brotli==1.2.0
cachetools==5.5.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.73.0
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
docker==7.1.0
entrypoints==0.4
executing==2.2.1
fastapi==0.120.4
ffmpy==0.6.4
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
gradio==5.49.1
gradio_client==1.13.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
groovy==0.1.2
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==22.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huey==2.5.4
huggingface-hub==1.0.1
idna==3.11
importlib_metadata==7.2.1
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mlflow==3.6.0
mlflow-skinny==3.6.0
mlflow-tracing==3.6.0
narwhals==2.10.1
nest-asyncio==1.6.0
numexpr @ file:///home/task_176216560327529/conda-bld/numexpr_1762165677686/work
numpy==1.26.4
openai==2.7.1
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
packaging==24.2
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==15.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.3
pyzmq==27.1.0
querystring-parser==1.2.4
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
ruff==0.14.3
safehttpx==0.1.7
scikit-learn==1.7.2
scipy==1.16.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.49.3
threadpoolctl==3.6.0
tomlkit==0.12.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0
